test(#2436): move negative pin into regression_1741h_tests.rs (file-size ratchet)
Folds the two roborev-1729 negative-pin tests (corrupt row_size VInt fails
CLOSED; uncorrupted control) as a nested row_size_bound_negative_pin module
inside the ALREADY-registered regression_1741h_tests.rs (already gated
#[cfg(all(test, feature = "write-support"))], 376 lines, well under the
1500-line test threshold) instead of adding a new top-level mod to
row_decoder/mod.rs. mod.rs is now byte-identical to its origin/main baseline
(no growth) -- the standalone file from the prior commit is removed.
